Add files via upload
Kindly check this executable file by downloading it and please don't worry it is not a virus file.
The executable file is a GUI based application of Dictionary it saves words if you wanted it to in a file called words.xls and since I am new to programming I am doing all the basic training in Python Language.
Please share this and let me know of any changes I should perform in this program. 
Happy coding.